Refactor: Move List Types Out for Testing
Need to wrap some testing as the default for list types does not work
with my server.

Split them out to their own submodule and added some basic conversion
testing.
