+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "reqwest_dav"
-version = "0.1.8"
+version = "0.1.9"
 edition = "2021"
 description = "An async webdav client with tokio and reqwest"
 keywords = ["http", "request", "client", "webdav"]
